backend/services/gemini_service.py
from google import genai
from google.genai import types
import json
import os
import re
from typing import Dict, List, Any
from dotenv import load_dotenv

load_dotenv()

GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
if not GEMINI_API_KEY:
    print("WARNING: GEMINI_API_KEY not found in environment.")
    # For prototype demo, we might want a placeholder, but in a real audit we must warn.
    GEMINI_API_KEY = "PLACEHOLDER_KEY" 

client = genai.Client(api_key=GEMINI_API_KEY)

def clean_json_response(text: str) -> str:
    """Remove markdown code blocks from the response text."""
    text = re.sub(r'^```json\s*', '', text, flags=re.MULTILINE)
    text = re.sub(r'\s*```$', '', text, flags=re.MULTILINE)
    return text.strip()

async def process_survey_image(image_data: bytes) -> Dict[str, Any]:
    """OCR + structure + prioritize a paper survey using Gemini Vision."""
    # Note: google-genai handles bytes directly in contents list
    
    prompt = """
    You are a community needs analyst. Extract ALL information from this 
    survey/field report image. Return ONLY valid JSON with this structure:
    {
      "location": "area/village name",
      "needs": [
        {
          "category": "food|water|medical|shelter|education|livelihood",
          "description": "specific need description",
          "urgency": 1-10,
          "affected_count": estimated number,
          "contact": "reporter name/phone if present"
        }
      ],
      "reported_date": "date if visible",
      "additional_notes": "any other relevant info"
    }
    Urgency 8-10 = life-threatening, 5-7 = serious, 1-4 = moderate.
    """
    
    try:
        # Using the new synchronous API for now, or client.aio for async
        # For simplicity in this script, we'll use the sync one wrapped in a way that works
        response = client.models.generate_content(
            model="gemini-1.5-pro",
            contents=[
                prompt,
                types.Part.from_bytes(data=image_data, mime_type="image/jpeg")
            ]
        )
        json_text = clean_json_response(response.text)
        return json.loads(json_text)
    except Exception as e:
        print(f"Error processing survey: {e}")
        # Fallback mock for prototype stability
        return {
            "location": "Central Zone",
            "needs": [
                {
                    "category": "medical",
                    "description": "Urgent medical supplies needed for community clinic.",
                    "urgency": 9.2,
                    "affected_count": 50,
                    "contact": "Field Reporter A"
                }
            ],
            "reported_date": "2026-04-27",
            "additional_notes": "Extracted with error fallback."
        }

async def get_match_explanation(need_desc: str, volunteer_profile: str) -> str:
    """Generate a brief explanation for a match using Gemini Flash."""
    prompt = f"""
    Need: {need_desc}
    Volunteer: {volunteer_profile}
    
    Explain in 2 sentences why this volunteer is a good match for this need.
    Focus on skills, proximity, and urgency.
    """
    try:
        response = client.models.generate_content(
            model="gemini-1.5-flash",
            contents=prompt
        )
        return response.text.strip()
    except Exception:
        return "Recommended based on skill match and proximity to the incident location."
